requirements.txt
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
datasets
ragas
supabase
pydantic
streamlit